Refactor device login tests (#2959)
* Persist CLI, PSContext credentials, handle Secure Strings

* Support PSCred/CLI auth inheritance in e2e

* Add testing for persistence and tenantID specification changes

* Refactor device login test

* Remove stub type

* Merge main

---------
