+++ e2etest/newe2e_runazcopy_stdout.go
@@ -264,3 +264,39 @@ func (a *AzCopyParsedLoginStatusStdout) Write(p []byte) (n int, err error) {
 	}
 	return a.AzCopyParsedStdout.Write(p)
 }
+
+var _ AzCopyStdout = &AzCopyInteractiveStdout{}
+
+// AzCopyInteractiveStdout is still a semi-raw stdout struct.
+type AzCopyInteractiveStdout struct {
+	Messages []string
+	asserter Asserter
+}
+
+// NewInteractiveWriter creates a new InteractiveWriter instance.
+func NewAzCopyInteractiveStdout(a Asserter) *AzCopyInteractiveStdout {
+	return &AzCopyInteractiveStdout{
+		asserter: a,
+	}
+}
+
+func (a *AzCopyInteractiveStdout) RawStdout() []string {
+	return a.Messages
+}
+
+func (a *AzCopyInteractiveStdout) Write(p []byte) (n int, err error) {
+	str := string(p)
+	lines := strings.Split(strings.TrimSuffix(str, "\n"), "\n")
+	n = len(p)
+
+	for _, v := range lines {
+		a.Messages = append(a.Messages, v)
+		a.asserter.Log(v)
+	}
+
+	return
+}
+
+func (a *AzCopyInteractiveStdout) String() string {
+	return strings.Join(a.RawStdout(), "\n")
+}
